requirements.txt
fastmcp>=3.1.0
imapclient>=3.0.0
html2text>=2024.2.26
python-dotenv>=1.0.0
uvicorn>=0.34.0
